rename test.ts to binding.ts
