+++ test/renderers/treemasterdetail.array.test.tsx
@@ -13,7 +13,7 @@ import {
   renderIntoDocument,
   scryRenderedDOMElementsWithClass,
   scryRenderedDOMElementsWithTag
-} from '../helpers/test';
+} from '../helpers/binding';
 import { Provider } from '../../src/common/binding';
 
 test.beforeEach(t => {
