+++ test/renderers/treemasterdetail.object.test.tsx
@@ -9,7 +9,7 @@ import {
   findRenderedDOMElementWithClass,
   findRenderedDOMElementWithTag,
   renderIntoDocument
-} from '../helpers/test';
+} from '../helpers/binding';
 import { Provider } from '../../src/common/binding';
 
 test.beforeEach(t => {
